requirements.txt
Flask
Flask-SQLAlchemy
Flask-Marshmallow
marshmallow
marshmallow-sqlalchemy
mysql-connector-python
